requirements.txt
aider-chat==0.13.0
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.7.1
APScheduler==3.9.1.post1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bitarray==2.8.1
blinker==1.6.2
cachetools==5.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
ConfigArgParse==1.7
cytoolz==0.12.2
discord-py-interactions==5.10.0
discord-typings==0.7.0
discord-ui==5.1.6
discord.py==2.3.2
diskcache==5.6.1
Django==4.2.5
EasyProcess==1.1
emoji==2.8.0
entrypoint2==1.1
environ==1.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
exceptiongroup==1.1.2
Flask==2.3.2
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
hexbytes==0.3.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
MailTm==0.0.81
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
networkx==3.1
numpy==1.24.3
openai==0.27.6
outcome==1.2.0
packaging==23.1
parsimonious==0.9.0
pefile==2023.2.7
Pillow==10.0.0
pluggy==1.2.0
prompt-toolkit==3.0.38
protobuf==4.24.0
psutil==5.9.5
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.18.0
pygame==2.5.1
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pystray==0.19.4
pyTelegramBotAPI==4.12.0
pytest==7.3.1
python-dotenv==1.0.0
python-telegram-bot==20.4
pytweening==1.0.7
pytz==2023.3
pyunormalize==15.0.0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.30.0
rfc3986==1.5.0
rich==13.3.5
rlp==3.0.0
rpds-py==0.9.2
scipy==1.10.1
selenium==4.10.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
sqlparse==0.4.4
telegram==0.0.1
threaded==4.1.0
tiktoken==0.4.0
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
undetected-chromedriver2==3.5.1
urllib3==2.0.2
watchdog==3.0.0
wcwidth==0.2.6
web3==6.8.0
websockets==11.0.3
Werkzeug==2.3.6
win10toast==0.9
WMI==1.5.1
wsproto==1.2.0
yarl==1.9.2
